Move ThumbnailModule into its own package
